use context for ladder data. fix player page ladder position doesnt work. import table columns instead of taking up space
