+++ client/src/App.tsx
@@ -1,26 +1,30 @@
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import { Navbar } from "./components/Navbar";
 import { Homepage } from "./components/Homepage";
-import { SoloLadder } from "./components/SoloLadder";
+import { PlayerLadderData, SoloLadder } from "./components/SoloLadder";
 import { GemHuntRecords } from "./components/GemHuntRecords";
 import { About } from "./components/About";
 import { Contact } from "./components/Contact";
 import { SignIn } from "./components/SignIn";
 import { SignUp } from "./components/SignUp";
 import { PageNotFound } from "./components/PageNotFound";
-import {useState } from 'react';
+import {useEffect, useState, createContext } from 'react';
 import { PlayerInfo } from "./components/PlayerInfo";
-
-
-
+import { getLadderData } from "./utils/utils";
 export const BASE_ROUTE = process.env.NODE_ENV === 'production' ? '' : 'http://localhost:3001'
+export const LadderData = createContext<PlayerLadderData[]>([]);
 
 function App() {
-  const [loggedInUser, setLoggedInUser] = useState<string>('')
+  const [ladderData, setLadderData] = useState<PlayerLadderData[]>([]);
+  useEffect(() => {
+    getLadderData(setLadderData);
+  }, []);
+
   return (
     <BrowserRouter>
     <Navbar/>
     <div className="fixed -z-10 w-screen h-screen bg-cover bg-fixed bg-[url('/public/MIU_1.jpg')] overflow-y-hidden" />
+    <LadderData.Provider value={ladderData}>
       <Routes>
         <Route path="/" element={<Homepage/>} />
         <Route path="/register" element={<SignUp/>} />
@@ -33,6 +37,7 @@ function App() {
         <Route path="/404" element={<PageNotFound/>} />
         <Route path="*" element={<Navigate to="/404"/>}/>
       </Routes>
+      </LadderData.Provider>
     </BrowserRouter>
   );
 }
